% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_betaexpg.R
\name{order_betaexpg}
\alias{order_betaexpg}
\title{Random Sampling of Order Statistics from a Beta Exponential G Distribution}
\usage{
order_betaexpg(size, spec, lambda, a, b, k, n)
}
\arguments{
\item{size}{numeric, represents the size of the sample.}

\item{spec}{character, represents an specific distribution. Possible values "norm", "exp","lnorm".}

\item{lambda}{numeric, represents the first shape parameter. Default value is 1.}

\item{a}{numeric, represents the second shape parameter. Default value is 1.}

\item{b}{numeric, represents the third shape parameter. Default value is 1.}

\item{k}{numeric, represents the Kth smallest value from a sample.}

\item{n}{numeric, represents the size of the sample to compute the order statistic from.}
}
\description{
\code{order_betaexpg} is used to obtain a random sample of order statistics from a Beta Exponential G Distribution.
}
\examples{
library(orders)
order_betaexpg(10,"exp",1,1,1,1,50)
order_betaexpg(10,"norm",1,1,1,1,50)
order_betaexpg(10,"lnorm",1,1,1,1,50)
}
\references{
Gentle, J, Computational Statistics, First Edition. Springer - Verlag, 2009.

Naradajah, S. and Rocha, R. (2016) Newdistns: An R Package for New Families of Distributions, Journal of Statiscal Software.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>.
}
