% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_eepg.R
\name{order_eepg}
\alias{order_eepg}
\title{Random Sampling of Order Statistics from a Exponentiated Exponential Poisson G Distribution}
\usage{
order_eepg(size, spec, lambda, a, k, n)
}
\arguments{
\item{size}{numeric, represents the size of the sample.}

\item{spec}{character, represents an specific distribution. Possible values "norm", "exp","lnorm".}

\item{lambda}{numeric, represents a scale parameter. Default value is 1.}

\item{a}{numeric, represents the shape parameter. Default value is 1.}

\item{k}{numeric, represents the error bound of the approximation of P(t). Default values is 0.01.}

\item{n}{numeric, represents the error bound of the approximation of P(t). Default values is 0.01.}
}
\description{
\code{order_eepg} is used to obtain a random sample of order statistics from a Exponentiated Exponential Poisson G Distribution.
}
\examples{
library(orders)
order_eepg(10,"exp",1,1,1,50)
order_eepg(10,"norm",1,1,1,50)
order_eepg(10,"norm",1,1,1,50)
}
\references{
Gentle, J, Computational Statistics, First Edition. Springer - Verlag, 2009.

Naradajah, S. and Rocha, R. (2016) Newdistns: An R Package for New Families of Distributions, Journal of Statiscal Software.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>.
}
