% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_zip.R
\name{order_zip}
\alias{order_zip}
\title{Random Sampling of k-th Order Statistics from a Zero Inflated Poisson Distribution}
\usage{
order_zip(size, k, mu, sigma, n, p = 0.5, alpha = 0.05, ...)
}
\arguments{
\item{size}{numeric, represents the size of the sample.}

\item{k}{numeric, represents the K-th smallest value from a sample.}

\item{mu}{numeric, represents the location parameter values.}

\item{sigma}{numeric, represents scale parameter values.}

\item{n}{numeric, represents the size of the sample to compute the order statistic from.}

\item{p}{numeric, represents the 100p percentile for the distribution of the K-th order statistic. Default value is population median, p = 0.5.}

\item{alpha}{numeric, (1 - alpha) represents the confidence of an interval for the population percentile p of the distribution of the k-th order statistic. Default value is 0.05.}

\item{...}{represents others parameters of a Zero Inflated Poisson distribution.}
}
\value{
A list with a random sample of order statistics from a Zero Inflated Poisson Distribution, the value of its join probability density function evaluated in the random sample and
and an approximate (1 - alpha) confidence interval for the population percentile p of the k-th order statistic.
}
\description{
\code{order_zip} is used to obtain a random sample of the k-th order statistic from a Zero Inflated Poisson distribution and some associated quantities of interest.
}
\examples{
library(orders)
# A sample of size 20 of the 5-th order statistics from a Zero Inflated Poisson Distribution
#order_zip(size=10,k=5,mu=5,sigma=0.1,n=30,p=0.5)
}
\references{
Gentle, J, Computational Statistics, First Edition. Springer - Verlag, 2009.

Ribgy, R. and Stasinopoulos, M. (2005) Generalized Additive Models for Location Scale and Shape, Journal of the Royal Statistical Society. Applied Statistics, Series C.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>.
}
